Move vespa IDs search tests into separate file for readability
